+++ src/main/java/com/example/demo/src/scholarship_comment/model/PatchScholarshipCommentReq.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PatchScholarshipCommentReq {
-    private Long scholarship_comment_idx;
+    private Integer scholarship_comment_idx;
     private String scholarship_comment_content;
 }
