+++ src/main/java/com/example/demo/src/support_comment/SupportCommentDao.java
@@ -23,14 +23,14 @@ public class SupportCommentDao {
     /**
      * 댓글 조회 API
      */
-    public List<GetSupportCommentRes> getSupportComment(Long support_idx) {
+    public List<GetSupportCommentRes> getSupportComment(Integer support_idx) {
         String getSupportCommentQuery = "select * from Support_Comment where support_idx = ? and support_comment_status = 'Y'";
-        Long getSupportCommentParams = support_idx;
+        Integer getSupportCommentParams = support_idx;
         return this.jdbcTemplate.query(getSupportCommentQuery,
                 (rs, rowNum) -> new GetSupportCommentRes(
-                        rs.getLong("support_comment_idx"),
-                        rs.getLong("support_idx"),
-                        rs.getLong("user_idx"),
+                        rs.getInt("support_comment_idx"),
+                        rs.getInt("support_idx"),
+                        rs.getInt("user_idx"),
                         rs.getString("support_comment_content"),
                         rs.getString("support_comment_updateAt")),
                 getSupportCommentParams); // 해당 닉네임을 갖는 모든 User 정보를 얻기 위해 jdbcTemplate 함수(Query, 객체 매핑 정보, Params)의 결과 반환
@@ -39,13 +39,13 @@ public class SupportCommentDao {
     /**
      * 댓글 작성 API
      */
-    public long createSupportComment(PostSupportCommentReq postSupportCommentReq) {
+    public Integer createSupportComment(PostSupportCommentReq postSupportCommentReq) {
         String createSupportCommentQuery = "insert into Support_Comment(support_idx, user_idx, support_comment_content) VALUES (?,?,?)"; // 실행될 동적 쿼리문
         Object[] createSupportCommentParams = new Object[]{postSupportCommentReq.getSupport_idx(), postSupportCommentReq.getUser_idx(), postSupportCommentReq.getSupport_comment_content()}; // 동적 쿼리의 ?부분에 주입될 값
         this.jdbcTemplate.update(createSupportCommentQuery, createSupportCommentParams);
 
         String lastInsertIdQuery = "select last_insert_id()"; // 가장 마지막에 삽입된(생성된) id값은 가져온다.
-        return this.jdbcTemplate.queryForObject(lastInsertIdQuery, long.class); // 해당 쿼리문의 결과 마지막으로 삽인된 유저의 BoardIdx번호를 반환한다.
+        return this.jdbcTemplate.queryForObject(lastInsertIdQuery, Integer.class); // 해당 쿼리문의 결과 마지막으로 삽인된 유저의 BoardIdx번호를 반환한다.
     }
